feat: transition governance deployment to Forge (#162)
When implementing the governance deployment, we hit a few roadblocks, and introduced hardhat for this step. While it good thing that it kept us moving forward, it caused this repo to get way more complicated and introduced additional cognitive load to the process. This PR removes that and reimplements the deployment in the existing paradigm by doing the following:

Deploy the GovernanceFactory in a solc-0.8 script.
Create a proposal and checks following the existing pattern.
Create a new cgp:execute command, which uses a foundry script to trigger execution. This has the added benefit of automatically verifying any contracts created as part of that transaction (i.e., all contracts that the Factory creates).
Replace the celocli command in the cgp:pass script with the cgp:execute command.
Move the merkle tree related scripts into a bin/ command to keep the repo structure.
This PR adds a mento-core-2.3.0 submodule, which currently points to a branch containing all the latest changes to the core contracts, but in the end, it will be pinned to the v2.3.0 tag. That version is used for the Governance MU.

Additionally, I have prepared #163, which makes it easy to verify contracts on celoscan as long as they were verified on sourcify. More details in that PR.
---------

Co-authored-by: philbow61 <philip.raetsch@gmail.com>